QUALITY: Remove abstract behaviour of push multiple nodes
LifoFrontier was not adding this in correct order. According to
lecture slides, last in should NOT be the right-most node, but
the left-most node

    	      x
	     / \
	    a   b

Add a to the end of the stack. The appendContentsOf: method
would push [a,b] to the LIFO stack. This means that b would
be traversed next (pop the top of the stack - which is b)
but we really need to traverse a instead