SPEED: Remove two for loops when finding blank tile
Instead of using two for loops to find the exact index of the blank
tile, instead we will use a flat map and the index of the blank
tile.

Example:

Show that the position is (1,2) using a flat map

1,2,3,4,5
6,7,0,8,9

becomes

1,2,3,4,5,6,7,0,8,9

What is the index of 0? It is 7
What is the width of the matrix? It is 5

For the row number, ask how many times the index goes into the width?

7 / 5 is = 1

For the col number, ask the remainder of how many times the index
goes into the width?

7 % 5 is = 2

Thus the position of 0 is (1,2)

Q.E.D