NEW: Add evaluation functions
Evaluation functions differ depending on the search being used.

This should be its own data structure to be plugged into the right
search method's frontier (i.e., the frontier evaluates differently
depending on the search method).

Rename the HeuristicFrontier to an EvaluatedFrontier to better
reflect the name of how the frontier does its work