QUALITY: Remove optional State for a Node
All nodes should have a state. If they don't then something
is seriously wrong and there should be a fatalError instead.

It doesn't make sense that we're unwrapping optional node state's
everywhere just to support isValidAction. If we apply an invalid
action to a state, let's just throw a fatal error instead because
this would be equivalent to undefined behaviour, e.g., insert
in an array at index -1 -- doesn't make sense!