FIX: Fix hash value implementation for large states
A large state would cause an Int overflow. Consider a state with
a sequence:

[1,2,3,...,44,45,0]

The value that would be reduced here is too large for an Int, and
as the implementation needs to be a Int, not a UInt, we have to
redo the implementation. I sourced a hashValue implementation
from a sequence from here:

     http://codereview.stackexchange.com/a/111573